users: tool unsupported on OpenBSD
  - utmpx not supported on OpenBSD

  - add src/uu/users/src/platform directory and platform/mod.rs for conditional compilation
    according to target_os
  - platform/openbsd.rs: implementation on OpenBSD (unsupported tool)
  - platform/unix.rs: implementation on other OS
  - src/uu/users/src/users.rs: use platform module for uucore::main function
